Fixed warnings in my files (Greg)

+++ Dev/StaticSheepNeoBuildSystem/game/SheepEngine/Components/controllers/light patterns/CLightPatternController.cpp
@@ -174,7 +174,7 @@ namespace Framework
 
       for (int i = 0; i < 4; ++i)
       {
-        int currWeight = currWeights[i] * NUMOFBLOCKS;
+        int currWeight = (int)currWeights[i] * NUMOFBLOCKS;
         for (int j = 0; j < currWeight; ++j)
         {
           blocks = blocks | currBlock;
